+++ frontend/src/modules/Datasets/components/DatasetData.js
@@ -1,0 +1,25 @@
+import PropTypes from 'prop-types';
+import React from 'react';
+import { Box } from '@mui/material';
+import { DatasetTables } from './DatasetTables';
+import { DatasetFolders } from './DatasetFolders';
+
+export const DatasetData = (props) => {
+  const { dataset, isAdmin } = props;
+
+  return (
+    <Box>
+      <Box>
+        <DatasetTables dataset={dataset} isAdmin={isAdmin} />
+      </Box>
+      <Box sx={{ mt: 3 }}>
+        <DatasetFolders dataset={dataset} isAdmin={isAdmin} />
+      </Box>
+    </Box>
+  );
+};
+
+DatasetData.propTypes = {
+  dataset: PropTypes.object.isRequired,
+  isAdmin: PropTypes.bool.isRequired
+};
