+++ frontend/src/modules/Datasets/components/index.js
@@ -1,2 +1,10 @@
 export * from './DatasetAWSActions';
 export * from './DatasetConsoleAccess';
+export * from './DatasetData';
+export * from './DatasetFolders';
+export * from './DatasetGovernance';
+export * from './DatasetListItem';
+export * from './DatasetOverview';
+export * from './DatasetStartCrawlerModal';
+export * from './DatasetTables';
+export * from './DatasetUpload';
